@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 # This spec was generated by rspec-rails when you ran the scaffold generator.
 # It demonstrates how one might use RSpec to specify the controller code that
@@ -18,146 +18,141 @@ require 'spec_helper'
 # Message expectations are only used when there is no simpler way to specify
 # that an instance is receiving a specific message.
 
-describe ContactsController do
+RSpec.describe ContactsController, type: :controller do
 
   # This should return the minimal set of attributes required to create a valid
   # Contact. As you add validations to Contact, be sure to
-  # update the return value of this method accordingly.
-  def valid_attributes
-    {}
-  end
+  # adjust the attributes here as well.
+  let(:valid_attributes) {
+    { name: 'foo' }
+  }
+
+  let(:invalid_attributes) {
+    { name: '' }
+  }
 
   # This should return the minimal set of values that should be in the session
   # in order to pass any filters (e.g. authentication) defined in
   # ContactsController. Be sure to keep this updated too.
-  def valid_session
-    {}
-  end
+  let(:valid_session) { {} }
 
-  describe "GET index" do
+  describe "GET #index" do
     it "assigns all contacts as @contacts" do
       contact = Contact.create! valid_attributes
-      get :index, {}, valid_session
-      assigns(:contacts).should eq([contact])
+      get :index, params: {}, session: valid_session
+      expect(assigns(:contacts)).to eq([contact])
     end
   end
 
-  describe "GET show" do
+  describe "GET #show" do
     it "assigns the requested contact as @contact" do
       contact = Contact.create! valid_attributes
-      get :show, {:id => contact.to_param}, valid_session
-      assigns(:contact).should eq(contact)
+      get :show, params: {id: contact.to_param}, session: valid_session
+      expect(assigns(:contact)).to eq(contact)
     end
   end
 
-  describe "GET new" do
+  describe "GET #new" do
     it "assigns a new contact as @contact" do
-      get :new, {}, valid_session
-      assigns(:contact).should be_a_new(Contact)
+      get :new, params: {}, session: valid_session
+      expect(assigns(:contact)).to be_a_new(Contact)
     end
   end
 
-  describe "GET edit" do
+  describe "GET #edit" do
     it "assigns the requested contact as @contact" do
       contact = Contact.create! valid_attributes
-      get :edit, {:id => contact.to_param}, valid_session
-      assigns(:contact).should eq(contact)
+      get :edit, params: {id: contact.to_param}, session: valid_session
+      expect(assigns(:contact)).to eq(contact)
     end
   end
 
-  describe "POST create" do
-    describe "with valid params" do
+  describe "POST #create" do
+    context "with valid params" do
       it "creates a new Contact" do
         expect {
-          post :create, {:contact => valid_attributes}, valid_session
+          post :create, params: {contact: valid_attributes}, session: valid_session
         }.to change(Contact, :count).by(1)
       end
 
       it "assigns a newly created contact as @contact" do
-        post :create, {:contact => valid_attributes}, valid_session
-        assigns(:contact).should be_a(Contact)
-        assigns(:contact).should be_persisted
+        post :create, params: {contact: valid_attributes}, session: valid_session
+        expect(assigns(:contact)).to be_a(Contact)
+        expect(assigns(:contact)).to be_persisted
       end
 
       it "redirects to the created contact" do
-        post :create, {:contact => valid_attributes}, valid_session
-        response.should redirect_to(Contact.last)
+        post :create, params: {contact: valid_attributes}, session: valid_session
+        expect(response).to redirect_to(Contact.last)
       end
     end
 
-    describe "with invalid params" do
+    context "with invalid params" do
       it "assigns a newly created but unsaved contact as @contact" do
-        # Trigger the behavior that occurs when invalid params are submitted
-        Contact.any_instance.stub(:save).and_return(false)
-        post :create, {:contact => {}}, valid_session
-        assigns(:contact).should be_a_new(Contact)
+        post :create, params: {contact: invalid_attributes}, session: valid_session
+        expect(assigns(:contact)).to be_a_new(Contact)
       end
 
       it "re-renders the 'new' template" do
-        # Trigger the behavior that occurs when invalid params are submitted
-        Contact.any_instance.stub(:save).and_return(false)
-        post :create, {:contact => {}}, valid_session
-        response.should render_template("new")
+        post :create, params: {contact: invalid_attributes}, session: valid_session
+        expect(response).to render_template("new")
       end
     end
   end
 
-  describe "PUT update" do
-    describe "with valid params" do
+  describe "PUT #update" do
+    context "with valid params" do
+      let(:new_attributes) {
+        { name: 'bar' }
+      }
+
       it "updates the requested contact" do
         contact = Contact.create! valid_attributes
-        # Assuming there are no other contacts in the database, this
-        # specifies that the Contact created on the previous line
-        # receives the :update_attributes message with whatever params are
-        # submitted in the request.
-        Contact.any_instance.should_receive(:update).with({'name' => 'updated'})
-        put :update, {:id => contact.to_param, :contact => {'name' => 'updated'}}, valid_session
+        put :update, params: {id: contact.to_param, contact: new_attributes}, session: valid_session
+        contact.reload
+        expect(contact.name).to eq(new_attributes[:name])
       end
 
       it "assigns the requested contact as @contact" do
         contact = Contact.create! valid_attributes
-        put :update, {:id => contact.to_param, :contact => valid_attributes}, valid_session
-        assigns(:contact).should eq(contact)
+        put :update, params: {id: contact.to_param, contact: valid_attributes}, session: valid_session
+        expect(assigns(:contact)).to eq(contact)
       end
 
       it "redirects to the contact" do
         contact = Contact.create! valid_attributes
-        put :update, {:id => contact.to_param, :contact => valid_attributes}, valid_session
-        response.should redirect_to(contact)
+        put :update, params: {id: contact.to_param, contact: valid_attributes}, session: valid_session
+        expect(response).to redirect_to(contact)
       end
     end
 
-    describe "with invalid params" do
+    context "with invalid params" do
       it "assigns the contact as @contact" do
         contact = Contact.create! valid_attributes
-        # Trigger the behavior that occurs when invalid params are submitted
-        Contact.any_instance.stub(:save).and_return(false)
-        put :update, {:id => contact.to_param, :contact => {}}, valid_session
-        assigns(:contact).should eq(contact)
+        put :update, params: {id: contact.to_param, contact: invalid_attributes}, session: valid_session
+        expect(assigns(:contact)).to eq(contact)
       end
 
       it "re-renders the 'edit' template" do
         contact = Contact.create! valid_attributes
-        # Trigger the behavior that occurs when invalid params are submitted
-        Contact.any_instance.stub(:save).and_return(false)
-        put :update, {:id => contact.to_param, :contact => {}}, valid_session
-        response.should render_template("edit")
+        put :update, params: {id: contact.to_param, contact: invalid_attributes}, session: valid_session
+        expect(response).to render_template("edit")
       end
     end
   end
 
-  describe "DELETE destroy" do
+  describe "DELETE #destroy" do
     it "destroys the requested contact" do
       contact = Contact.create! valid_attributes
       expect {
-        delete :destroy, {:id => contact.to_param}, valid_session
+        delete :destroy, params: {id: contact.to_param}, session: valid_session
       }.to change(Contact, :count).by(-1)
     end
 
     it "redirects to the contacts list" do
       contact = Contact.create! valid_attributes
-      delete :destroy, {:id => contact.to_param}, valid_session
-      response.should redirect_to(contacts_url)
+      delete :destroy, params: {id: contact.to_param}, session: valid_session
+      expect(response).to redirect_to(contacts_url)
     end
   end
 
