@@ -1,11 +1,10 @@
-require 'spec_helper'
+require 'rails_helper'
 
-describe "Contacts" do
+RSpec.describe "Contacts", type: :request do
   describe "GET /contacts" do
     it "works! (now write some real specs)" do
-      # Run the generator again with the --webrat flag if you want to use webrat methods/matchers
       get contacts_path
-      response.status.should be(200)
+      expect(response).to have_http_status(200)
     end
   end
 end
