@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe "contacts/show" do
   before(:each) do
@@ -15,11 +15,11 @@ describe "contacts/show" do
   it "renders attributes in <p>" do
     render
     # Run the generator again with the --webrat flag if you want to use webrat matchers
-    rendered.should match(/Name/)
-    rendered.should match(/Address/)
-    rendered.should match(/City/)
-    rendered.should match(/State/)
-    rendered.should match(/Postal Code/)
-    rendered.should match(/Phone/)
+    expect(rendered).to match(/Name/)
+    expect(rendered).to match(/Address/)
+    expect(rendered).to match(/City/)
+    expect(rendered).to match(/State/)
+    expect(rendered).to match(/Postal Code/)
+    expect(rendered).to match(/Phone/)
   end
 end
